refactor: add unit-remove module
