+++ core-commonui/unit-remove/src/commonMain/kotlin/com/github/diegoberaldin/raccoonforlemmy/unit/remove/di/Utils.kt
@@ -1,0 +1,8 @@
+package com.github.diegoberaldin.raccoonforlemmy.unit.remove.di
+
+import com.github.diegoberaldin.raccoonforlemmy.unit.remove.RemoveMviModel
+
+expect fun getRemoveViewModel(
+    postId: Int? = null,
+    commentId: Int? = null,
+): RemoveMviModel
