+++ settings.gradle.kts
@@ -50,5 +50,6 @@ include(":core-commonui:unit-selectcommunity")
 include(":core-commonui:unit-drawer")
 include(":core-commonui:unit-communityinfo")
 include(":core-commonui:unit-instanceinfo")
+include(":core-commonui:unit-remove")
 include(":core-commonui:detailopener-api")
 include(":core-commonui:detailopener-impl")
